Filters Added

+++ KursovaApp/Classes/University.cs
@@ -9,12 +9,12 @@ public class University : IDescriptionable
     public string Country { get; set;}
     public int StudentsCount { get; set;}
     public List<StudyField> StudyFields { get; set;}
-    public double Price { get; set;}
+    public int Price { get; set;}
     public string Description { get; set; }
 
     public string PhotoPath { get; set; }
 
-    public University(string _Name, string _City, string _Country, int _StudentsCount, List<StudyField> _StudyField, double _Price, string _Description, string _PhotoPath)
+    public University(string _Name, string _City, string _Country, int _StudentsCount, List<StudyField> _StudyField, int _Price, string _Description, string _PhotoPath)
     {
         Name = _Name;
         City = _City;
@@ -25,7 +25,7 @@ public class University : IDescriptionable
         Description = _Description;
         PhotoPath = _PhotoPath;
     }
-    public University(string _Name, string _City, string _Country, int _StudentsCount, double _Price)
+    public University(string _Name, string _City, string _Country, int _StudentsCount, int _Price)
     {
         Name = _Name;
         City = _City;
@@ -33,7 +33,7 @@ public class University : IDescriptionable
         StudentsCount = _StudentsCount;
         Price = _Price;
     }
-    public University(string _Name, string _City, string _Country, int _StudentsCount, double _Price, string _Description)
+    public University(string _Name, string _City, string _Country, int _StudentsCount, int _Price, string _Description)
     {
         Name = _Name;
         City = _City;
